add admin

+++ azure-user.sh
@@ -16,7 +16,7 @@ fi
 #create func
 create()
 {
-    password=Donthackme(987
+    password="Donthackme(987"
     domain=alexvo19gmail.onmicrosoft.com
     displayname=$1
     subscription=$2
@@ -59,20 +59,46 @@ delete()
 
     #checks if user exits
     user=$(az ad user list --query [].userPrincipalName | grep $userprincipalname)
-
-    #conditional 
+    
+    #conditional to see if user exis
     if [ -z $user ]
     then
      echo "user doesn't exist"
      exit 1
     fi
 
+    #check if user that is to be deleted is admin
+    checkuser=$(az role assignment list --include-classic-administrators \
+    --query "[?roleDefinitionId=='NA(classic admin role)'].principalName" \
+    | grep -E $userprincipalname)
+
+    if [ -n "$checkuser" ]
+     then
+     echo "can't delete another admin"
+     exit 1
+    fi
+
+
     az ad user delete --upn-or-object-id $userprincipalname
 }
 
 
-#runs different function based on argument given
 
+
+#must be admin to 
+username=$(az account show --query user.name)
+check=$(az role assignment list --include-classic-administrators \
+    --query "[?roleDefinitionId=='NA(classic admin role)'].principalName" \
+    | grep -E $username)
+
+if [ -z "$check" ]
+then
+  echo "not an admin"
+  exit 1
+fi
+
+
+#runs different function based on argument given
 case "$commands" in
   "create")
     username=$2
